move nets to flow matcher.
